refactor: Wave C hygiene — render race fix, main.js split, zero warnings
Implements Wave C of the v2 evaluation's Consolidate & Harden plan:

- Add a render-generation token to features/diagrams.js: both
  processMermaidDiagrams and reRenderMermaidDiagrams re-check a shared
  generation counter after every await and stop mutating DOM/panzoom
  state once superseded, killing the rapid-tab-switch race that leaked
  panzoom instances and could interleave mermaid theme config with an
  in-flight render. Two regression tests simulate the race with a
  deferred mermaid.render.
- Decompose main.js from 977 to 505 lines: extract
  features/app-commands.js (palette registry), features/version-check.js
  (version label + GitHub poll; constants stay in main.js for
  sync-version.js), features/drag-drop.js, features/keyboard.js (global
  keydown + Esc chain), and platform/ios-wiring.js. Pure moves, no
  behavior change; unused imports and DOM refs pruned.
- Fix all 49 lint warnings (optional catch binding for unused catch
  params, dead renderer workspaceRoot removed, stale eslint-disable
  directives and stale vendor/ ignore dropped) and gate the lint script
  at --max-warnings=0 so warnings now fail CI.
- Object.seal(state) in core/state.js: field values stay mutable but the
  shared-state shape can no longer grow ad hoc.

Renderer coverage instrumentation stays deferred (documented): it
requires migrating all 31 suites off the global-eval test harness.

Tests: 482 passing (+2), typecheck clean, lint 0 errors/0 warnings
(enforced), build green.
