+++ docs/project-modernization/README.md
@@ -39,6 +39,7 @@ three-lens evaluation of the app at v0.0.82 and a phased roadmap.
 | 2026-07-14 | [brainstorm-evaluation-v2-post-modernization](2026-07-14-brainstorm-evaluation-v2-post-modernization.md) | **Evaluation v2** — independent re-audit of v0.0.154: v0.0.82 findings scorecard (all fixed except coverage thresholds), verdict that the design/UX layer is ~50% shipped (tokens unadopted, emoji icons + harsh dark theme remain, Comments/Annotate/Notes collision), new security findings (unsigned Win/Linux auto-update, `npm install` in release lanes, admin-PAT push), and a "Consolidate & Harden" next-cycle plan (Waves A–D) |
 | 2026-07-14 | [tasks-session-23-wave-a-hardening](2026-07-14-tasks-session-23-wave-a-hardening.md)                     | **Wave A implemented** — auto-update gated to signed platforms (Win/Linux fall back to a Releases-page flow), `npm ci` in release lanes, SW caches only `response.ok` navigations, workspace-root containment for relative links, macOS drops the redundant GitHub-API version poll, all Actions SHA-pinned, honest `coverageThreshold` on `desktop/main.js`; ADMIN_PAT scoping documented as a user action                              |
 | 2026-07-14 | [tasks-session-24-wave-b-ux-consolidation](2026-07-14-tasks-session-24-wave-b-ux-consolidation.md)       | **Wave B implemented** — SVG icon set replaces emoji, Comments/Annotate/Notes collision resolved, overflow "⋮" menu real at all widths (toolbar: Contents/Split/Annotate/Present/Search), dark-theme accent split with AA contrast, targeted motion tokens, focus traps on all modals, completed shortcut sheet, URL-open loading state, first-run hierarchy                                                                             |
+| 2026-07-14 | [tasks-session-25-wave-c-hygiene](2026-07-14-tasks-session-25-wave-c-hygiene.md)                         | **Wave C implemented** — render-generation token kills the tab-switch race (+2 regression tests), `main.js` decomposed 977→505 lines into five modules, all 49 lint warnings fixed with `--max-warnings=0` now enforced, `core/state.js` sealed against shape drift; renderer coverage instrumentation deferred with rationale                                                                                                           |
 | 2026-06-14 | [handoff-next-wave](2026-06-14-handoff-next-wave.md)                                                     | **Resume point** — aggregated status, the remaining Phase 3/4 options, and the project gotchas. Start here after a break.                                                                                                                                                                                                                                                                                                                |
 
 > **Resuming after a break?** Read **[handoff-next-wave](2026-06-14-handoff-next-wave.md)** first — it's the durable index of what's done, what's next, and the gotchas.
