+++ package.json
@@ -15,7 +15,7 @@
     "test:ci": "jest --ci --coverage --maxWorkers=2",
     "test:report": "jest --verbose --json --outputFile=test-results.json",
     "typecheck": "tsc --noEmit -p tsconfig.json",
-    "lint": "eslint .",
+    "lint": "eslint . --max-warnings=0",
     "lint:fix": "eslint . --fix",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
